Create test_steam.py
Allows testing against official steam files, assuming that a 'GAMES' environment variables is set.

It is possible to use registry (on windows) and common paths to find steam, but that's prone to error, and the testing env needs to be run by someone, so they can set it up themselves.